% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord_convert}
\alias{coord_convert}
\title{Coordinate System Conversion}
\usage{
coord_convert(lnglat, from, to)
}
\arguments{
\item{lnglat}{location in the form of longitude + latitude, seperated by ','}

\item{from, to}{one of "WGS-84", "GCJ-02" or "BD-09". Can be abbrivated.}
}
\description{
Coordinate System Conversion
}
