% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_location}
\alias{get_location}
\title{Get Location from input arguments}
\usage{
get_location(lnglat = NULL, latlng = NULL, longitude = NULL,
  latitude = NULL)
}
\description{
Get Location from input arguments
}
