% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi.R
\name{poi.amap}
\alias{poi.amap}
\title{POI Search using amap}
\usage{
poi.amap(keywords, city = NULL, types = NULL, citylimit = FALSE,
  children = 0, offset = 20, page = 1, extensions = c("base",
  "all"), sign = FALSE, output = c("JSON", "XML"), opts = list(),
  headers = list())
}
\arguments{
\item{keywords}{address and place keywords}

\item{city}{city}
}
\description{
POI Search using amap
}
\details{
Reference: https://lbs.amap.com/api/webservice/guide/api/search
}
