% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi.R
\name{poi.baidu}
\alias{poi.baidu}
\title{Baidu Map POI Search}
\usage{
poi.baidu(query, region, tag = NULL, city_limit = NULL,
  output = c("json", "xml"), scope = c("1", "2"), filter = NULL,
  coord_type = 3, ret_coordtype = c("bd09ll", "gcj02ll", "bd09mc"),
  page_size = 10, page_num = 0, sign = TRUE, opts = list(),
  headers = list())
}
\arguments{
\item{query}{address and place keywords}

\item{region}{city or province name. Refer to http://wiki.lbsyun.baidu.com/cms/citycode/BaiduMap_cityCode_1102.zip}

\item{tag}{category e.g. 酒店、美食. Refer to http://lbsyun.baidu.com/index.php?title=lbscloud/poitags for more detail}

\item{city_limit}{TRUE/FALSE, whether to restrict result to the city}

\item{output}{output type: json or xml}

\item{scope}{basic (1) or detailed (2) POI information}

\item{filter}{filter result}

\item{coord_type}{request coord_type, possible values include
\{1: 'wgs84ll', 2: 'gcj02ll', 3: 'bd09ll' (default), 4: 'bd09ll'\}}

\item{ret_coordtype}{response coord_type, possible value 'gcj02ll'. Refer to http://lbsyun.baidu.com/index.php?title=coordinate}
}
\description{
Baidu Map POI Search
}
\details{
Ref: http://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-placeapi
}
