% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_location}
\alias{reverse_location}
\title{Reverse Location}
\usage{
reverse_location(location)
}
\arguments{
\item{location}{character string of longitude latitude in the form of "lng,lat"}
}
\value{
reversed location in the form of "lat,lng"
}
\description{
Reverse Location
}
