% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_location}
\alias{split_location}
\title{Split Location}
\usage{
split_location(location, reverse = FALSE, simplify = TRUE)
}
\arguments{
\item{location}{character string of longitude latitude in the form of "lng,lat"}
}
\value{
list of longitude, latitude pairs in the form of c(lng, lat)
}
\description{
Split Location
}
